+++ AngularJSAuthentication.API/Models/TemplateTask.cs
@@ -23,8 +23,8 @@ namespace AngularJSAuthentication.API.Models
             Active = true;
         }
 
-        public int TemplateId { get; set; }
-        public int TaskId { get; set; }
+        public long TemplateId { get; set; }
+        public long TaskId { get; set; }
         public string Stage { get; set; }
         public string Environment { get; set; }
         public string Domain { get; set; }
